Split up map rendering code
