+++ src/handlers/walk.ts
@@ -7,9 +7,9 @@ import {
   WalkReplyServerPacket,
 } from 'eolib';
 import type { Client } from '../client';
-import { CharacterWalkAnimation } from '../character';
 import { getPrevCoords } from '../utils/get-prev-coords';
 import { inRange } from '../utils/range';
+import { CharacterWalkAnimation } from '../render/character-walk';
 
 function handleWalkPlayer(client: Client, reader: EoReader) {
   const packet = WalkPlayerServerPacket.deserialize(reader);
